+++ code/game/bg_panimate.c
@@ -756,7 +756,7 @@ qboolean BG_ParseAnimationFile(const char *filename)
 {
 	char		*text_p;
 	int			len;
-	int			i;
+	unsigned	i;
 	char		*token;
 	float		fps;
 	int			skip;
@@ -776,7 +776,7 @@ qboolean BG_ParseAnimationFile(const char *filename)
 		{
 			return qfalse;
 		}
-		if ( len >= sizeof( BGPAFtext ) - 1 )
+		if ( len >= (int)sizeof( BGPAFtext ) - 1 )
 		{
 			//Com_Printf( "File %s too long\n", filename );
 			return qfalse;
